Updated Binary class to hold a vector of vectors of capstone::cs_insn in a field called 'instructions'. Added the function 'disassemble' to the Binary class to populate this field using capstone::disassemble from capstone.rs. Began to add the function 'analyze' to the Binary class, but it is left unfinished for now.
Modified capstone::disassemble to return a Result<Vec<Vec<cs_insn>>,
cs_err> in order to preserve the instructions recursively disassembled.
A vector of vectors may be unnecessary and will be evaluated.

Modified the main function in main.rs to properly handle the new
functionality for disassembly in the Binary class.