Updated Binary class to store a vector of BasicBlocks instead a vector of vectors of cs_insn.
Added the function `instructions` to return a vector of all instructions
from each basic block combined. This should make it easier for logic
that just needs raw data.

Updated the BasicBlock class to properly split blocks. Also changed the
`self` parameter to be a reference to avoid borrowing errors in
BasicBlock methods.