Lots of changes due to the fact nothing would work unless a large amount of changes where made.
Added struct Edge to represent edges. Currently contains an entry and
exit, but will be expanded in the future when edge types are needed.

Changed the typed of 'edges' in CFG to use the new Edge class.
Changed the 'new' function to include a copy of the text segment in
order to ensure only valid addresses are searched.
Added three new methods: 'get_successors', 'get_predecessors', and
'get_block'. These assist in CFG creation, but can be used at any time.

The simple iteration over basic blocks is now used to create a temporary
edge set in order to build a more complete set. Afterwards, we check for
calls and create a vector to use as a stack of return addresses. Using
(the probably wrongly named) 'dfs' helper function, we traverse the
graph until a return is found, recursively traversing addition calls
found, until a return is found and an edge can be added from the block
containing the return instruction to the address on the return stack.

This will potentially create many edges into and out of functions. Thus,
in order to process this new graph, a return stack similar what was used
to create this graph must be used when processing it.