Added the Graph trait to make DFS generalizable. Also added the Vertex trait in an attempt to make the Graph trait more generalizable. Did not attempt to generalize edges though.
Added more functions that I believe will be useful to the Graph trait,
as well as default implementations. It is clear that there is actually a
mistake in this commit, because only 'get_edges' and 'get_vertices'
should be implementation specific. Both 'add_edge' and 'add_vertex'
should have default implementations. 'get_vertices' should return a
reference to the underlying vertex set. This will be updated in the next
commit, probably with changes to CFG to accommodate this.

Lifetimes caused issues in DominatorTree and DJGraph, so we are no
longer storing references to CFGs and instead will consume them.

Added the field 'start' to DominatorTree to serve as the root. This is
needed to construct a dominator tree. Implemented Graph for both
DominatorTree and DJGraph, since it will be needed for DFS, which is
also required to construct both.