Changed the type of the 'vertices' field in CFG, DominatorTree, and DJGraph to be an owned vector in order to removed unneeded complexity in the implementations of each.
Changed the return type of the default implementations of
'get_predecessors' and 'get_successors' in the Graph trait. HashSets
were showing to be too limited (or I not knowledgeable enough) to
implement a function to determine connected components.

Since call is no longer considered an unconditional jump, a portion of
the 'detect_edges' function in CFG has been commented out for testing.
Results from this commit will be compared to results in previous commits
to avoid manual checking of edges.

Added the beginning of the function 'connected_components', which will
take an input graph and create a vector of collections (currently a
HashSet) of vertices that are contained in each connected component.

The main purpose of this incomplete commit is to test changes in output
after no longer considering call instructions to be unconditional jumps.