This commit has unfinished code.
This is the beginning of changes made to implement Sreedhar, et al.'s
loop detection algorithm.

Implemented SpanningTree to represent a spanning tree that classifies
edges as sp-tree, sp-back, sp-forward, or sp-cross edges. SpanningTree
can also find a path between vertices, if one exists. Identifying
sp-back edges is critical for implementing Sreedhar, et al.'s algorithm.

To make spanning tree creation simpler, Ordering is now a struct that
contains both tree edges and ordering information. All uses of 'dfs'
have been updated to handle this change.

Added a helper function to the implementation of Edge to return an edge
with the same entry and exit, but with a different edge type. Also added
a new field, 'max_level', to DJGraph to make implementing Sreedhar, et
al.'s algorithm easier.

Began to implement Sreedhar, et al.'s algorithm as the method
'detect_loops' of DJGraph, but it is not complete. Tarjan's strongly
connected component (SCC) algorithm (or similar) must be implemented
first in order to detect the body of irreducible loops. A method for
collapsing loops into a single vertex must also be implemented before
this algorithm can be completed.

Temporarily commented out 'connected_components' because it will likely
need to be rewritten and it is not needed at the current time.