Loop detection still in progress. SCC and collapsing loops in SpanningTree will be required in order to properly detect loops.
Implemented 'reach_under' that is analogous to ReachUnder in Sreedhar,
et al.'s work. This algorithm finds all vertices, on the same level or
lower (higher in value) of the destination of an sp-back edge, that can
reach the source of said sp-back edge. Implemented 'path', similar to
the method in SpanningTree, to help with this.

The implementation of DJGraph::detect_loops in the last commit had
errors due to borrowing objects both immutably and mutably. Since
HashSet does not implement mutable iterators, we now loop over of vector
of Vec<HashSet> where each HashSet contains copies of all vertices on a
specific level. Added corresponding function calls and comments in
places where they are need, even if the corresponding function calls are
not yet implemented properly.

Discovered that collapsed loop bodies were not correctly produced if the
body only contained one vertex and that multiple edges that loop back to
a vertex were being added. Added necessary checks to resolve this.

Discovered that Tarjan's SCC algorithm modifies the state of many
objects, which appear to be globally accessible. In order to avoid
global variables, we would like to pass them as function parameters, but
it seems like there are enough to warrant a struct. This will likely be
implemented in one of the next few commits.

Changed the type of vector in the 'body' field of Loop. Generics can be
implemented later after the algorithms work. Also implemented
fmt::Display in order to easily print loop information.