Created CFG::compoenents to split a CFG into connected components and fixed a bug in DJGraph::detect_loops.
CFG::subgraph uses only DFS to find nodes and does not create a
connected graph. CFG::components attempts to return all connected
components of the CFG. It uses CFG::find_root to attempt to identify a
root, but may not find the correct root, and there may not be a single
root.

Attempted to fix a bug in DominatorTree where vertices without predecessors
would produce an empty list of dominators. The temporary fix involved
adding the starting node of the CFG to the list of dominators, but this
is probably incorrect and may be attributing to the low number of loops
detected.

Fixed a bug in DJGraph::path where infinite recursion was causing stack
overflows. The function no longer visits vertices it has already
visited.