Made several updates due to long processing time and memory usage of dominator tree creation.
After processing a 1.1 MB file for 6 hours, changed the 'vertices' field
of each class to be a hash map that stores predecessors and successors
for each vertex. This makes computation of 'get_predecessors' and
'get_successors' much faster, but causes memory usage to explode.

Changed all address/id related fields from u64 to u32 to attempt to cut
down on memory usage for vertex/edge sets over 10,000 in size. This has
not been tested for effectiveness as of this commit, however (as of
typing, 2 hours have passed on processing an ~500KB file with > 10,000
vertices and > 20,000 edges using ~6GB of memory while using u64 as
address/id field).